tendo_django/core/exceptions.py
# -*- coding: utf-8 -*-
from django.utils.translation import ugettext_lazy as _
from rest_framework import serializers
from rest_framework.exceptions import APIException
from rest_framework.views import exception_handler


def custom_exception_handler(exc, context):
    # Call REST framework's default exception handler first,
    # to get the standard error response.
    response = exception_handler(exc, context)
    # Now add the HTTP status code to the response.
    if response is not None:
        response.data['status_code'] = response.status_code

    return response

class InvalidAPIQuery(APIException):
    """Default Exception for Api Queries"""
    status_code = 400
    default_detail = 'An invalid query parameter was provided'
    default_code = 'invalid_query'