requirements.txt
django==3.2
djangorestframework==3.12.4
django-environ==0.4.5

psycopg2-binary
django-model-utils==3.1.2
